+++ Source/DlgSystem/Private/DlgManager.cpp
@@ -31,8 +31,7 @@ UDlgContext* UDlgManager::StartDialogueWithDefaultParticipants(UObject* WorldCon
 	}
 
 	// Create empty map of participants we need
-	TSet<FName> ParticipantSet;
-	Dialogue->GetAllParticipantNames(ParticipantSet);
+	TSet<FName> ParticipantSet = Dialogue->GetParticipantNames();
 	TArray<UObject*> Participants;
 
 	// Maps from Participant Name => Objects that have that participant name
@@ -43,7 +42,7 @@ UDlgContext* UDlgManager::StartDialogueWithDefaultParticipants(UObject* WorldCon
 	}
 
 	// Gather all objects that have our participant name
-	for (UObject* Participant : GetAllObjectsWithDialogueParticipantInterface(WorldContextObject))
+	for (UObject* Participant : GetObjectsWithDialogueParticipantInterface(WorldContextObject))
 	{
 		const FName ParticipantName = IDlgDialogueParticipant::Execute_GetParticipantName(Participant);
 		if (ObjectMap.Contains(ParticipantName))
@@ -285,7 +284,7 @@ TArray<TWeakObjectPtr<AActor>> UDlgManager::GetAllWeakActorsWithDialogueParticip
 	return Array;
 }
 
-TArray<UObject*> UDlgManager::GetAllObjectsWithDialogueParticipantInterface(UObject* WorldContextObject)
+TArray<UObject*> UDlgManager::GetObjectsWithDialogueParticipantInterface(UObject* WorldContextObject)
 {
 	TArray<UObject*> Array;
 	if (!WorldContextObject)
@@ -322,11 +321,11 @@ TArray<UObject*> UDlgManager::GetAllObjectsWithDialogueParticipantInterface(UObj
 	return Array;
 }
 
-TMap<FName, FDlgObjectsArray> UDlgManager::GetAllObjectsMapWithDialogueParticipantInterface(UObject* WorldContextObject)
+TMap<FName, FDlgObjectsArray> UDlgManager::GetObjectsMapWithDialogueParticipantInterface(UObject* WorldContextObject)
 {
 	// Maps from Participant Name => Objects that have that participant name
 	TMap<FName, FDlgObjectsArray> ObjectsMap;
-	for (UObject* Participant : GetAllObjectsWithDialogueParticipantInterface(WorldContextObject))
+	for (UObject* Participant : GetObjectsWithDialogueParticipantInterface(WorldContextObject))
 	{
 		const FName ParticipantName = IDlgDialogueParticipant::Execute_GetParticipantName(Participant);
 		if (ObjectsMap.Contains(ParticipantName))
@@ -446,92 +445,108 @@ TArray<UDlgDialogue*> UDlgManager::GetAllDialoguesForParticipantName(FName Parti
 	return DialoguesArray;
 }
 
-void UDlgManager::GetAllDialoguesParticipantNames(TArray<FName>& OutArray)
+TArray<FName> UDlgManager::GetDialoguesParticipantNames()
 {
 	TSet<FName> UniqueNames;
 	for (const UDlgDialogue* Dialogue : GetAllDialoguesFromMemory())
 	{
-		Dialogue->GetAllParticipantNames(UniqueNames);
+		UniqueNames.Append(Dialogue->GetParticipantNames());
 	}
 
-	FDlgHelper::AppendSortedSetToArray(UniqueNames, OutArray);
+	TArray<FName> Array;
+	FDlgHelper::AppendSortedSetToArray(UniqueNames, Array);
+	return Array;
 }
 
-void UDlgManager::GetAllDialoguesSpeakerStates(TArray<FName>& OutArray)
+TArray<FName> UDlgManager::GetDialoguesSpeakerStates()
 {
 	TSet<FName> UniqueNames;
 	for (const UDlgDialogue* Dialogue : GetAllDialoguesFromMemory())
 	{
-		Dialogue->GetAllSpeakerStates(UniqueNames);
+		UniqueNames.Append(Dialogue->GetSpeakerStates());
 	}
 
-	FDlgHelper::AppendSortedSetToArray(UniqueNames, OutArray);
+	TArray<FName> Array;
+	FDlgHelper::AppendSortedSetToArray(UniqueNames, Array);
+	return Array;
 }
 
-void UDlgManager::GetAllDialoguesIntNames(FName ParticipantName, TArray<FName>& OutArray)
+TArray<FName> UDlgManager::GetDialoguesParticipantIntNames(FName ParticipantName)
 {
 	TSet<FName> UniqueNames;
 	for (const UDlgDialogue* Dialogue : GetAllDialoguesFromMemory())
 	{
-		Dialogue->GetIntNames(ParticipantName, UniqueNames);
+		UniqueNames.Append(Dialogue->GetParticipantIntNames(ParticipantName));
 	}
 
-	FDlgHelper::AppendSortedSetToArray(UniqueNames, OutArray);
+	TArray<FName> Array;
+	FDlgHelper::AppendSortedSetToArray(UniqueNames, Array);
+	return Array;
 }
 
-void UDlgManager::GetAllDialoguesFloatNames(FName ParticipantName, TArray<FName>& OutArray)
+TArray<FName> UDlgManager::GetDialoguesParticipantFloatNames(FName ParticipantName)
 {
 	TSet<FName> UniqueNames;
 	for (const UDlgDialogue* Dialogue : GetAllDialoguesFromMemory())
 	{
-		Dialogue->GetFloatNames(ParticipantName, UniqueNames);
+		UniqueNames.Append(Dialogue->GetParticipantFloatNames(ParticipantName));
 	}
 
-	FDlgHelper::AppendSortedSetToArray(UniqueNames, OutArray);
+	TArray<FName> Array;
+	FDlgHelper::AppendSortedSetToArray(UniqueNames, Array);
+	return Array;
 }
 
-void UDlgManager::GetAllDialoguesBoolNames(FName ParticipantName, TArray<FName>& OutArray)
+TArray<FName> UDlgManager::GetDialoguesParticipantBoolNames(FName ParticipantName)
 {
 	TSet<FName> UniqueNames;
 	for (const UDlgDialogue* Dialogue : GetAllDialoguesFromMemory())
 	{
-		Dialogue->GetBoolNames(ParticipantName, UniqueNames);
+		UniqueNames.Append(Dialogue->GetParticipantBoolNames(ParticipantName));
 	}
 
-	FDlgHelper::AppendSortedSetToArray(UniqueNames, OutArray);
+	TArray<FName> Array;
+	FDlgHelper::AppendSortedSetToArray(UniqueNames, Array);
+	return Array;
 }
 
-void UDlgManager::GetAllDialoguesNameNames(FName ParticipantName, TArray<FName>& OutArray)
+TArray<FName> UDlgManager::GetDialoguesParticipantFNameNames(FName ParticipantName)
 {
 	TSet<FName> UniqueNames;
 	for (const UDlgDialogue* Dialogue : GetAllDialoguesFromMemory())
 	{
-		Dialogue->GetNameNames(ParticipantName, UniqueNames);
+		UniqueNames.Append(Dialogue->GetParticipantFNameNames(ParticipantName));
 	}
 
-	FDlgHelper::AppendSortedSetToArray(UniqueNames, OutArray);
+	TArray<FName> Array;
+	FDlgHelper::AppendSortedSetToArray(UniqueNames, Array);
+	return Array;
 }
 
-void UDlgManager::GetAllDialoguesConditionNames(FName ParticipantName, TArray<FName>& OutArray)
+TArray<FName> UDlgManager::GetDialoguesParticipantConditionNames(FName ParticipantName)
 {
 	TSet<FName> UniqueNames;
 	for (const UDlgDialogue* Dialogue : GetAllDialoguesFromMemory())
 	{
-		Dialogue->GetConditions(ParticipantName, UniqueNames);
+		UniqueNames.Append(Dialogue->GetParticipantConditionNames(ParticipantName));
 	}
 
-	FDlgHelper::AppendSortedSetToArray(UniqueNames, OutArray);
+	TArray<FName> Array;
+	FDlgHelper::AppendSortedSetToArray(UniqueNames, Array);
+	return Array;
 }
 
-void UDlgManager::GetAllDialoguesEventNames(FName ParticipantName, TArray<FName>& OutArray)
+TArray<FName> UDlgManager::GetDialoguesParticipantEventNames(FName ParticipantName)
 {
 	TSet<FName> UniqueNames;
 	for (const UDlgDialogue* Dialogue : GetAllDialoguesFromMemory())
 	{
-		Dialogue->GetEvents(ParticipantName, UniqueNames);
+		UniqueNames.Append(Dialogue->GetParticipantEventNames(ParticipantName));
 	}
 
-	FDlgHelper::AppendSortedSetToArray(UniqueNames, OutArray);
+	TArray<FName> Array;
+	FDlgHelper::AppendSortedSetToArray(UniqueNames, Array);
+	return Array;
 }
 
 bool UDlgManager::RegisterDialogueConsoleCommands()