+++ Source/DlgSystemEditor/Private/BlueprintNodes/DialogueK2Node_Select.cpp
@@ -516,34 +516,31 @@ void UDialogueK2Node_Select::GetPrintStringFunction(FName& FunctionName, UClass*
 bool UDialogueK2Node_Select::RefreshPinNames()
 {
 	// Stop anything if the blueprint is loading, this can happen because we now have reference to blueprint UClasses (reflection system) from the UDlgDialogue
-	if (!FDialogueBlueprintUtilities::IsBlueprintLoadedForGraphNode(this))
-	{
-		return false;
-	}
-
-	const FName ParticipantName = FDialogueBlueprintUtilities::GetParticipantNameFromNode(this);
+	static constexpr bool bBlueprintMustBeLoaded = true;
+	const FName ParticipantName = FDialogueBlueprintUtilities::GetParticipantNameFromNode(this, bBlueprintMustBeLoaded);
 	if (ParticipantName == NAME_None && VariableType != EDlgVariableType::SpeakerState)
 	{
 		return false;
 	}
+	// EDlgVariableType::SpeakerState does not need a Participant
 
 	TArray<FName> NewPinNames;
 	switch (VariableType)
 	{
 		case EDlgVariableType::Float:
-			UDlgManager::GetAllDialoguesFloatNames(ParticipantName, NewPinNames);
+			NewPinNames.Append(UDlgManager::GetDialoguesParticipantFloatNames(ParticipantName));
 			break;
 
 		case EDlgVariableType::Int:
-			UDlgManager::GetAllDialoguesIntNames(ParticipantName, NewPinNames);
+			NewPinNames.Append(UDlgManager::GetDialoguesParticipantIntNames(ParticipantName));
 			break;
 
 		case EDlgVariableType::Name:
-			UDlgManager::GetAllDialoguesNameNames(ParticipantName, NewPinNames);
+			NewPinNames.Append(UDlgManager::GetDialoguesParticipantFNameNames(ParticipantName));
 			break;
 
 		case EDlgVariableType::SpeakerState:
-			UDlgManager::GetAllDialoguesSpeakerStates(NewPinNames);
+			NewPinNames.Append(UDlgManager::GetDialoguesSpeakerStates());
 			break;
 
 		default: