+++ Source/DlgSystemEditor/Private/BlueprintNodes/DialogueK2Node_SwitchDialogueCallback.cpp
@@ -175,7 +175,8 @@ void UDialogueK2Node_SwitchDialogueCallback::AddPinToSwitchNode()
 // Begin own functions
 bool UDialogueK2Node_SwitchDialogueCallback::RefreshPinNames()
 {
-	const FName ParticipantName = FDialogueBlueprintUtilities::GetParticipantNameFromNode(this);
+	static constexpr bool bBlueprintMustBeLoaded = true;
+	const FName ParticipantName = FDialogueBlueprintUtilities::GetParticipantNameFromNode(this, bBlueprintMustBeLoaded);
 	if (ParticipantName == NAME_None)
 	{
 		return false;
@@ -185,27 +186,27 @@ bool UDialogueK2Node_SwitchDialogueCallback::RefreshPinNames()
 	switch (CallbackType)
 	{
 		case EDlgDialogueCallback::Event:
-			UDlgManager::GetAllDialoguesEventNames(ParticipantName, NewPinNames);
+			NewPinNames.Append(UDlgManager::GetDialoguesParticipantEventNames(ParticipantName));
 			break;
 
 		case EDlgDialogueCallback::Condition:
-			UDlgManager::GetAllDialoguesConditionNames(ParticipantName, NewPinNames);
+			NewPinNames.Append(UDlgManager::GetDialoguesParticipantConditionNames(ParticipantName));
 			break;
 
 		case EDlgDialogueCallback::FloatValue:
-			UDlgManager::GetAllDialoguesFloatNames(ParticipantName, NewPinNames);
+			NewPinNames.Append(UDlgManager::GetDialoguesParticipantFloatNames(ParticipantName));
 			break;
 
 		case EDlgDialogueCallback::IntValue:
-			UDlgManager::GetAllDialoguesIntNames(ParticipantName, NewPinNames);
+			NewPinNames.Append(UDlgManager::GetDialoguesParticipantIntNames(ParticipantName));
 			break;
 
 		case EDlgDialogueCallback::BoolValue:
-			UDlgManager::GetAllDialoguesBoolNames(ParticipantName, NewPinNames);
+			NewPinNames.Append(UDlgManager::GetDialoguesParticipantBoolNames(ParticipantName));
 			break;
 
 		case EDlgDialogueCallback::NameValue:
-			UDlgManager::GetAllDialoguesNameNames(ParticipantName, NewPinNames);
+			NewPinNames.Append(UDlgManager::GetDialoguesParticipantFNameNames(ParticipantName));
 			break;
 
 		default: