+++ Source/DlgSystemEditor/Private/DialogueEditor/DetailsPanel/DialogueSpeechSequenceEntry_Details.h
@@ -48,19 +48,15 @@ public:
 
 private:
 	/** Gets the ParticipantNames from all Dialogues. */
-	TArray<FName> GetAllDialoguesParticipantNames() const
+	TArray<FName> GetDialoguesParticipantNames() const
 	{
-		TArray<FName> OutArray;
-		UDlgManager::GetAllDialoguesParticipantNames(OutArray);
-		return OutArray;
+		return UDlgManager::GetDialoguesParticipantNames();
 	}
 
 	/** Gets the Speaker States from all Dialogues. */
-	TArray<FName> GetAllDialoguesSpeakerStates() const
+	TArray<FName> GetDialoguesSpeakerStates() const
 	{
-		TArray<FName> OutArray;
-		UDlgManager::GetAllDialoguesSpeakerStates(OutArray);
-		return OutArray;
+		return UDlgManager::GetDialoguesSpeakerStates();
 	}
 
 	/** Gets the current Dialogue Participant Names. */